+++ .gitignore
@@ -1,6 +1,3 @@
 .env
 node_modules
-dist/*.js
-dist/main.js
-dist/
 dist
