+++ package.json
@@ -18,9 +18,6 @@
   "author": "StarNumber12046",
   "license": "MIT",
   "dependencies": {
-    "@types/body-parser": "^1.19.2",
-    "@types/dotenv": "^8.2.0",
-    "@types/express": "^4.17.14",
     "body-parser": "^1.20.1",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
@@ -28,9 +25,12 @@
     "mongoose": "^6.6.5"
   },
   "devDependencies": {
+    "@types/body-parser": "^1.19.2",
+    "@types/express": "^4.17.14",
     "@typescript-eslint/eslint-plugin": "^5.39.0",
     "@typescript-eslint/parser": "^5.39.0",
     "eslint": "^8.24.0",
-    "jest": "^29.1.2"
+    "jest": "^29.1.2",
+    "typescript": "^4.8.4"
   }
 }
